- Printing alert if there is no file with specified pattern - Printing message after merge

+++ lib/merger/filemerger.rb
@@ -7,18 +7,22 @@ module Merger
       first_mask_files = Searcher.find_files_for_mask(config.masks.first)
       first_mask_files.each do |first_mask_file|
         file_name = File.basename(first_mask_file).to_s.chomp(config.masks.first)
-
         content = ""
         config.masks.each do |mask|
           file = File.dirname(first_mask_file) + "/" + file_name + mask
-          content += File.readlines(file).join() + "\n"
-          delete_file_if_needed(file, config)
+          if File.exist?(file)
+            content += File.readlines(file).join() + "\n"
+            delete_file_if_needed(file, config)
+          else
+            Poster.post_file_not_found(file)
+          end
         end
 
         new_file_name = File.dirname(first_mask_file) + "/" + file_name + config.result_mask
         File.open(new_file_name, "w") { |f| f.puts content }
         puts "❇️  File #{new_file_name} created"
       end
+      Poster.post_merge_finished
     end
 
     private
